1.修复ipaddress定义成int导致的一整天数据combine失败的bug 2。修复之前由于hashcode写的有问题导致的出现重复记录的BUG

+++ ServerComponent/OnlineUserPackage/src/main/java/com/huayu/metis/entry/RegisterLogEntry.java
@@ -39,7 +39,7 @@ public class RegisterLogEntry extends BaseLogEntry {
         this.appId = Integer.parseInt(values[2]);
         this.registerMode = Integer.parseInt(values[3]);
         this.terminalCode = Integer.parseInt(values[4]);
-        this.ipAddress = Integer.parseInt(values[5]);
+        this.ipAddress = Long.parseLong(values[5]);
         //转换注册日期
         Date regDate = sdf.parse(values[6]);
         this.registerTime = regDate.getTime();
